+++ android/src/main/java/techmind/studio/detect_clap_sound_flutter/DetectClapSoundFlutterPlugin.java
@@ -9,6 +9,9 @@ import android.os.Looper;
 import androidx.annotation.NonNull;
 
 import java.io.IOException;
+import java.util.LinkedList;
+import java.util.Map;
+import java.util.Queue;
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.embedding.engine.plugins.activity.ActivityAware;
@@ -19,6 +22,7 @@ import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 import techmind.studio.detect_clap_sound_flutter.constants.ArgumentConstants;
+import techmind.studio.detect_clap_sound_flutter.config.DetectConfig;
 import techmind.studio.detect_clap_sound_flutter.constants.MethodConstants;
 import techmind.studio.detect_clap_sound_flutter.permission.PermissionManager;
 import techmind.studio.detect_clap_sound_flutter.permission.PermissionResultCallback;
@@ -44,7 +48,9 @@ public class DetectClapSoundFlutterPlugin implements FlutterPlugin, ActivityAwar
 
     private Handler handler;
 
-    private int amplitudeThreshold = 18000;
+    private Queue<Integer> amplitudeQueue = new LinkedList<>();
+
+    private boolean detectedRecently = false;
 
     private boolean isRecording = false;
 
@@ -91,11 +97,14 @@ public class DetectClapSoundFlutterPlugin implements FlutterPlugin, ActivityAwar
                 result.success(isRecording);
                 break;
 
+            case MethodConstants.DISPOSE_RECORDING_METHOD:
+                disposeRecording(result);
+                break;
+
             default:
                 result.notImplemented();
                 break;
         }
-
     }
 
     @Override
@@ -115,7 +124,6 @@ public class DetectClapSoundFlutterPlugin implements FlutterPlugin, ActivityAwar
             permissionManager.setActivity(mActivity);
             activityBinding.addRequestPermissionsResultListener(pm);
         }
-
     }
 
     @Override
@@ -140,7 +148,6 @@ public class DetectClapSoundFlutterPlugin implements FlutterPlugin, ActivityAwar
 
         activityBinding = null;
         mActivity = null;
-
     }
 
 
@@ -158,26 +165,37 @@ public class DetectClapSoundFlutterPlugin implements FlutterPlugin, ActivityAwar
     /**
      * [startRecording]
      * To start recording.
-     * Listen and detect sound with [amplitudeThreshold].
      * Use [eventSink] to emit event to Flutter.
-     * Default [amplitudeThreshold] is 18000 for Clap Sound.
+     * Identifies spikes in sound amplitude, helping to detect specific sound events such as clapping or whistling.
      */
     private void startRecording(MethodCall call, MethodChannel.Result result) {
+        /// Get arguments.
+        String path = call.argument(ArgumentConstants.PATH_FILE);
+        if (path == null || path.isEmpty()) {
+            path = context.getExternalFilesDir(null).getAbsolutePath();
+        }
+
+        String fileName = call.argument(ArgumentConstants.FILE_NAME);
+        if (fileName == null || fileName.isEmpty()) {
+            fileName = "music";
+        }
+
+        /// Get detect config map.
+        final Map<String, Object> configMap = call.argument(ArgumentConstants.DETECT_CONFIG);
+
+        /// Define default config.
+        DetectConfig detectConfig = new DetectConfig(
+                getOrDefault(configMap, "threshold", 30000),
+                getOrDefault(configMap, "stableThreshold", 3000),
+                getOrDefault(configMap, "amplitudeSpikeThreshold", 28000),
+                getOrDefault(configMap, "monitorInterval", 50),
+                getOrDefault(configMap, "windowSize", 5)
+        );
+
         recorder = new MediaRecorder();
         recorder.setAudioSource(MediaRecorder.AudioSource.MIC);
         recorder.setOutputFormat(MediaRecorder.OutputFormat.THREE_GPP);
         recorder.setAudioEncoder(MediaRecorder.AudioEncoder.AMR_NB);
-
-        /// Get arguments.
-        String path = call.argument(ArgumentConstants.PATH_FILE);
-        if (path == null || path.isEmpty()) {
-            path = mActivity.getExternalFilesDir(null).getAbsolutePath();
-        }
-
-        String fileName = call.argument(ArgumentConstants.FILE_NAME);
-        if(fileName == null || fileName.isEmpty()){
-            fileName = "music";
-        }
 
         /// Set audio file path.
         audioFilePath = path + "/" + fileName + ".3gp";
@@ -190,21 +208,33 @@ public class DetectClapSoundFlutterPlugin implements FlutterPlugin, ActivityAwar
             isRecording = true;
 
             handler = new Handler(Looper.getMainLooper());
+
             runnable = new Runnable() {
                 @Override
                 public void run() {
 
+                    /// Get Max Ampl and add to [amplitudeQueue].
                     int amplitude = recorder.getMaxAmplitude();
+                    amplitudeQueue.add(amplitude);
 
-                    if (amplitude >= amplitudeThreshold) {
-
-                        if (eventSink != null) {
-
-                            eventSink.success(1);
-                        }
+                    if (amplitudeQueue.size() > detectConfig.getWindowSize()) {
+                        amplitudeQueue.poll();
                     }
 
-                    handler.postDelayed(this, 250);
+                    if (isClapDetected(detectConfig)) {
+
+                        if (eventSink != null) {
+                            eventSink.success(1);
+                        }
+
+                        detectedRecently = true;
+                        amplitudeQueue.clear();
+
+                    } else if (isAmplitudeStable(detectConfig)) {
+                        detectedRecently = false;
+                    }
+
+                    handler.postDelayed(this, detectConfig.getMonitorInterval());
                 }
             };
             handler.post(runnable);
@@ -213,6 +243,58 @@ public class DetectClapSoundFlutterPlugin implements FlutterPlugin, ActivityAwar
         } catch (IOException e) {
             result.error("RECORDING_FAILED", "Recording failed", e.getMessage());
         }
+    }
+
+
+    /**
+     * [isClapDetected]
+     * <p>
+     * To check the surge amplitude and rapid decrease.
+     * <p>
+     * [threshold]: Minimum amplitude threshold to identify a sound as possible clapping.
+     * <p>
+     * [stableThreshold]: Threshold to determine when the amplitude has returned to a stable level, helping to prevent repeated clap detection.
+     * <p>
+     * [amplitudeSpikeThreshold]: Threshold to determine the sudden change in amplitude (difference between largest and smallest amplitude), which characterizes applause.
+     * <p>
+     * [windowSize]: Number of amplitude samples in the test window.
+     * <p>
+     * [monitorInterval]: Interval between amplitude checks.
+     */
+    private boolean isClapDetected(DetectConfig config) {
+        if (amplitudeQueue.size() < config.getWindowSize() || detectedRecently) {
+            return false;
+        }
+
+        int maxAmplitudeInWindow = 0;
+        int minAmplitudeInWindow = Integer.MAX_VALUE;
+
+        for (int amplitude : amplitudeQueue) {
+            if (amplitude > maxAmplitudeInWindow) {
+                maxAmplitudeInWindow = amplitude;
+            }
+            if (amplitude < minAmplitudeInWindow) {
+                minAmplitudeInWindow = amplitude;
+            }
+        }
+
+        return maxAmplitudeInWindow > config.getThreshold() &&
+                minAmplitudeInWindow < config.getStableThreshold() &&
+                (maxAmplitudeInWindow - minAmplitudeInWindow) > config.getAmplitudeSpikeThreshold();
+    }
+
+
+    /**
+     * [isAmplitudeStable]
+     * To get Amplitude Stable
+     */
+    private boolean isAmplitudeStable(DetectConfig config) {
+        for (int amplitude : amplitudeQueue) {
+            if (amplitude > config.getStableThreshold()) {
+                return false;
+            }
+        }
+        return true;
     }
 
 
@@ -241,5 +323,30 @@ public class DetectClapSoundFlutterPlugin implements FlutterPlugin, ActivityAwar
             recorder = null;
             isRecording = false;
         }
+    }
+
+    /**
+     * [disposeRecording]
+     * To clear record and stream listen
+     */
+    private void disposeRecording(MethodChannel.Result result) {
+        if (recorder != null) {
+            stopRecordingManually();
+            result.success("DISPOSE RECORDING SUCCESS");
+        } else {
+            result.error("NO_RECORDING", "No ongoing recording found", null);
+        }
+    }
+
+
+    /**
+     * [getOrDefault]
+     * Generic method to get a value from a map or return a default value if the key is not found
+     */
+    private <T> T getOrDefault(Map<String, Object> map, String key, T defaultValue) {
+        if (map != null && map.containsKey(key) && map.get(key) != null) {
+            return (T) map.get(key);
+        }
+        return defaultValue;
     }
 }