+++ android/src/main/java/techmind/studio/detect_clap_sound_flutter/constants/ArgumentConstants.java
@@ -5,4 +5,6 @@ public class ArgumentConstants {
     public static final String PATH_FILE = "path_file";
 
     public static final String FILE_NAME = "file_name";
+
+    public static final String DETECT_CONFIG = "detect_config";
 }