+++ lib/detect_clap_sound_flutter.dart
@@ -5,3 +5,4 @@ export 'src/constant/argument_constants.dart';
 export 'src/detect_clap_sound_flutter_method_channel.dart';
 export 'src/detect_clap_sound_flutter_platform_interface.dart';
 export 'src/detect_clap_sound_flutter.dart';
+export '/src/config/detect_config.dart';