+++ test/detect_clap_sound_flutter_test.dart
@@ -1,18 +1,16 @@
+import 'package:detect_clap_sound_flutter/detect_clap_sound_flutter.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:detect_clap_sound_flutter/src/detect_clap_sound_flutter.dart';
-import 'package:detect_clap_sound_flutter/src/detect_clap_sound_flutter_platform_interface.dart';
-import 'package:detect_clap_sound_flutter/src/detect_clap_sound_flutter_method_channel.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 class MockDetectClapSoundFlutterPlatform with MockPlatformInterfaceMixin implements DetectClapSoundFlutterPlatform {
   @override
-  Future<bool?> getStatusPermission() => Future.value(false);
+  Future<bool?> hasPermission() => Future.value(false);
 
   @override
   Future<bool?> requestPermission() => Future.value(false);
 
   @override
-  Future<String?> startRecording({String? fileName, String? path}) => Future.value(null);
+  Future<String?> startRecording({String? fileName, String? path, DetectConfig? config}) => Future.value(null);
 
   @override
   Future<String?> stopRecording() => Future.value(null);
@@ -23,7 +21,10 @@ class MockDetectClapSoundFlutterPlatform with MockPlatformInterfaceMixin impleme
   }
 
   @override
-  Future<bool?> getStatusRecording() => Future.value(false);
+  Future<bool?> isRecording() => Future.value(false);
+
+  @override
+  void dispose() {}
 }
 
 void main() {
@@ -38,6 +39,6 @@ void main() {
     MockDetectClapSoundFlutterPlatform fakePlatform = MockDetectClapSoundFlutterPlatform();
     DetectClapSoundFlutterPlatform.instance = fakePlatform;
 
-    expect(detectClapSoundFlutterPlugin.getStatusPermission, true);
+    expect(detectClapSoundFlutterPlugin.hasPermission, true);
   });
 }