+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 1.1.1
 
-* TODO: Describe initial release.
+* Initial release.
+* Implementations for Android.
+* Example.
+* Requires Flutter Version >= 3.19.0.
+* Requires Dart Version >= 3.3.0.